Separate handler package from event package
